requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==22.3.0
boto==2.49.0
boto3==1.21.42
botocore==1.24.42
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==4.0.4
django-rest-swagger==2.2.0
django-ses==3.0.1
djangorestframework==3.13.1
filelock==3.0.12
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
openapi-codec==1.3.2
pathspec==0.9.0
pipenv==2021.5.29
platformdirs==2.2.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
requests==2.27.1
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.1
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.7.2
virtualenv-clone==0.5.6

